More work towards v3.2-b2
Removed lots of work-arounds for RRF not having its own Linux task
Bug fix: Added missing "Starting" item to the MachineStatus enumeration
Bug fix: M112/M999 were executed out-of-order when read from files
Bug fix: Sometimes in print files codes invoking macro files could crash DCS
Bug fix: Aborted macro files did not cancel codes properly

+++ src/DuetControlServer/Program.cs
@@ -195,22 +195,37 @@ namespace DuetControlServer
                 }
             }
 
-            // Execute runonce.g if it is present
+            // Execute runonce.g after config.g if it is present
             string runOnceFile = await FilePath.ToPhysicalAsync(FilePath.RunOnceFile, FileDirectory.System);
             if (File.Exists(runOnceFile))
             {
-                using (Macro macro = new Macro(FilePath.RunOnceFile, runOnceFile, DuetAPI.CodeChannel.Trigger))
+                do
                 {
-                    await macro.FinishAsync();
+                    using (await Model.Provider.AccessReadOnlyAsync())
+                    {
+                        if (Model.Provider.Get.State.Status != MachineStatus.Starting)
+                        {
+                            break;
+                        }
+                    }
                 }
+                while (!CancellationToken.IsCancellationRequested);
 
-                try
+                if (!CancellationToken.IsCancellationRequested)
                 {
-                    File.Delete(runOnceFile);
-                }
-                catch (Exception e)
-                {
-                    await Model.Provider.Output(MessageType.Error, $"Failed to delete {FilePath.RunOnceFile}: {e.Message}");
+                    using (Macro macro = new Macro(FilePath.RunOnceFile, runOnceFile, DuetAPI.CodeChannel.Trigger))
+                    {
+                        await macro.FinishAsync();
+                    }
+
+                    try
+                    {
+                        File.Delete(runOnceFile);
+                    }
+                    catch (Exception e)
+                    {
+                        await Model.Provider.Output(MessageType.Error, $"Failed to delete {FilePath.RunOnceFile}: {e.Message}");
+                    }
                 }
             }
